requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Authentication & Security
PyJWT==2.8.0
cryptography==41.0.7
argon2-cffi==23.1.0
pyotp==2.9.0
python-multipart==0.0.6

# Database
mysql-connector-python==8.2.0
sqlalchemy==2.0.23
alembic==1.13.0

# Caching & Sessions
redis==5.0.1
hiredis==2.2.3

# HTTP & API
httpx==0.25.2
requests==2.31.0

# Data Validation & Serialization
python-jose[cryptography]==3.3.0
email-validator==2.1.0

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0
httpx==0.25.2

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1